[lint package refactor] Create a file for check_parallel
